+++ src/modules/catalog/migrations/index.ts
@@ -1,1 +1,0 @@
-export * from './1758000100000-create-products';