+++ src/modules/identity/migrations/index.ts
@@ -1,1 +1,0 @@
-export * from './1758000000000-create-users';